# [PM-20424] [BEEEP] browser client webpack optimisations