--- a/apps/browser/webpack.config.js
+++ b/apps/browser/webpack.config.js
@@ -76,11 +76,14 @@ const moduleRules = [
   },
   {
     test: /\.[cm]?js$/,
+    exclude: /node_modules\/(?!@bitwarden)/,
     use: [
       {
         loader: "babel-loader",
         options: {
           configFile: "../../babel.config.json",
+          cacheDirectory: ENV === "development",
+          compact: ENV !== "development",
         },
       },
     ],
@@ -88,6 +91,9 @@ const moduleRules = [
   {
     test: /\.[jt]sx?$/,
     loader: "@ngtools/webpack",
+    options: {
+      jitMode: ENV === "development",
+    },
   },
   {
     test: /argon2(-simd)?\.wasm$/,
@@ -157,6 +163,7 @@ const plugins = [
     tsConfigPath: "tsconfig.json",
     entryModule: "src/popup/app.module#AppModule",
     sourceMap: true,
+    jitMode: ENV === "development",
   }),
   new webpack.ProvidePlugin({
     process: "process/browser.js",
@@ -204,6 +211,20 @@ const mainConfig = {
     "encrypt-worker": "../../libs/common/src/key-management/crypto/services/encrypt.worker.ts",
     "content/send-on-installed-message": "./src/vault/content/send-on-installed-message.ts",
     "content/send-popup-open-message": "./src/vault/content/send-popup-open-message.ts",
+  },
+  cache:
+    ENV !== "development"
+      ? false
+      : {
+          type: "filesystem",
+          name: "main-cache",
+          cacheDirectory: path.resolve(__dirname, "../../node_modules/.cache/webpack-browser-main"),
+          buildDependencies: {
+            config: [__filename],
+          },
+        },
+  snapshot: {
+    unmanagedPaths: [path.resolve(__dirname, "../../node_modules/@bitwarden/")],
   },
   optimization: {
     minimize: ENV !== "development",
@@ -263,6 +284,7 @@ const mainConfig = {
       fs: false,
       path: require.resolve("path-browserify"),
     },
+    cache: true,
   },
   output: {
     filename: "[name].js",
@@ -357,6 +379,23 @@ if (manifestVersion == 2) {
       ],
       noParse: /argon2(-simd)?\.wasm$/,
     },
+    cache:
+      ENV !== "development"
+        ? false
+        : {
+            type: "filesystem",
+            name: "background-cache",
+            cacheDirectory: path.resolve(
+              __dirname,
+              "../../node_modules/.cache/webpack-browser-background",
+            ),
+            buildDependencies: {
+              config: [__filename],
+            },
+          },
+    snapshot: {
+      unmanagedPaths: [path.resolve(__dirname, "../../node_modules/@bitwarden/")],
+    },
     experiments: {
       asyncWebAssembly: true,
     },
@@ -369,6 +408,7 @@ if (manifestVersion == 2) {
         fs: false,
         path: require.resolve("path-browserify"),
       },
+      cache: true,
     },
     dependencies: ["main"],
     plugins: [...requiredPlugins],