## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-20424

## 📔 Objective

Speeds up Browser build for development by leveraging caching and turning off optimisations.

### Script: `build:watch`
- On M3 Pro with NODE_OPTIONS="--max-old-space-size=8192"
- Measured using `/usr/bin/time -l npm run build:watch`

|Script| Before                                                   | After                                                            |
|---|----------------------------------------------------------|------------------------------------------------------------------|
|First run, no cache| 45 seconds main, 30 seconds background, 5.2GB max memory | 25 seconds main, 31 seconds background, 5.2GB max memory  |
|Subsequent run with caching, minimal or no changes to source code| (the same as no cache)                                   | 16 seconds main, 1 second background, 3.9GB max memory |

#### Conclusion: 
- First run (no cache) run is up to 17 seconds (24%) faster.
- Subsequent (cached) run is up to 56 seconds (76%) faster and requires up to 1.3GB less memory.

No GH Actions difference - as expected.

## 📸 Screenshots

<!-- Required for any UI changes; delete if not applicable. Use fixed width images for better display. -->

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or informed the documentation team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire! -->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry that's not quite a confirmed issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes