@addisonbeck 2025-04-22
This seems redundant with an upcoming initiative to implement Nx across the monorepo. One of the use cases for Nx is cached builds across the entire monorepo dependency graph, including what is being done here.

Can you speak on this at all? Is there a reason to have both caches?