@mzieniukbw 2025-04-23
In spirit of incremental improvement, i don't see any harm having this in the meantime, while Nx is being adopted to browser client. 
When the time comes in to use Nx (i am assuming weeks-months), we can then remove the changes introduced in this PR.